+++ example/MultiplatformExample/shared/src/commonMain/kotlin/com/example/ktorfittest/StarWarsApi.kt
@@ -1,5 +1,6 @@
 package com.example.ktorfittest
 
+import de.jensklingenberg.ktorfit.Call
 import de.jensklingenberg.ktorfit.http.GET
 import de.jensklingenberg.ktorfit.http.Path
 import kotlinx.coroutines.flow.Flow
@@ -10,4 +11,7 @@ interface StarWarsApi {
 
     @GET("people/{id}/")
     fun getPeopleByIdFlowResponse(@Path("id") peopleId: Int): Flow<Person>
+
+    @GET("people/{id}/")
+    fun getPeopleByIdCallResponse(@Path("id") peopleId: Int): Call<Person>
 }