+++ ktorfit-lib/src/jvmTest/kotlin/TestEngine.kt
@@ -23,7 +23,7 @@ open class TestEngine() : HttpClientEngineBase("ktor-mock") {
 
     @InternalAPI
     override suspend fun execute(data: HttpRequestData): HttpResponseData {
-        val con = createCallContext(data.executionContext)
+        val coroutineContext1 = createCallContext(data.executionContext)
         getRequestData(data)
         return HttpResponseData(
             HttpStatusCode.Accepted,
@@ -31,7 +31,7 @@ open class TestEngine() : HttpClientEngineBase("ktor-mock") {
             Headers.Empty,
             HttpProtocolVersion.HTTP_2_0,
             "",
-            con
+            coroutineContext1
         )
     }