+++ sandbox/src/linuxX64Main/kotlin/LinuxMain.kt
@@ -1,6 +1,6 @@
 import com.example.api.JsonPlaceHolderApi
 import de.jensklingenberg.ktorfit.Ktorfit
-import de.jensklingenberg.ktorfit.adapter.FlowResponseConverter
+import de.jensklingenberg.ktorfit.converter.builtin.FlowResponseConverter
 import de.jensklingenberg.ktorfit.create
 import io.ktor.client.*
 import kotlinx.coroutines.runBlocking