+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.0-beta11
+========================================
+
+- you can now use ResponseConverter in combination with suspend functions. Implement the SuspendResponseConverter
+- KtorfitCallResponseConverter and FlowResponseConverter moved to de.jensklingenberg.ktorfit.converter.builtin
+
+
 1.0.0-beta10
 ========================================
 (18.08.2022)